style: Update S-dot styling & preview css
