+++ src/app/admin/write/page.tsx
@@ -3,84 +3,12 @@
 import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
-import { styled } from 'styled-components';
 import { tagAPI } from '@/api/tag';
 import { postAPI } from '@/api/post';
 import WithAuth from '@/components/hocs/WithAuth';
 import { Editor } from '@/components/markdown/Editor';
 import { Tag } from '@/types/Tag';
-
-const S = {
-  container: styled.div`
-    ${({ theme }) =>
-      theme.MIXINS.flexBox('column', 'flex-start', 'flex-start')};
-    height: 100%;
-    width: 100%;
-  `,
-  heading: styled.h1`
-    font-weight: bold;
-    font-size: 32px;
-    width: 100%;
-    text-align: center;
-    color: ${({ theme }) => theme.colors.text['222222']};
-    margin-top: 12px;
-  `,
-  section: styled.section`
-    width: 100%;
-    margin-top: 48px;
-    ${({ theme }) =>
-      theme.MIXINS.flexBox('column', 'flex-start', 'flex-start')};
-  `,
-  form: styled.form`
-    width: 100%;
-    margin-top: 12px;
-    padding: 20px;
-    
-    ${({ theme }) =>
-      theme.MIXINS.flexBox('column', 'flex-start', 'flex-start')};
-  `,
-  select: styled.select`
-    width: 200px;
-    height: 40px;
-    padding-left: 10px;
-    color: ${({ theme }) => theme.colors.text['222222']};
-    font-weight: 500;
-    border: 1px solid ${({ theme }) => theme.colors.text['222222']};
-  `,
-  input: styled.div`
-    width: 100%;
-    height: 40px;
-    margin-top: 12px;
-
-    input {
-      padding-left: 10px;
-      width: 100%;
-      height: 100%;
-      border: 1px solid ${({ theme }) => theme.colors.text['222222']};
-    }
-  `,
-  editorWrap: styled.div`
-    width: 100%;
-    margin-top: 12px;
-  `,
-  action: styled.input`
-    margin-top: 12px;
-    ${({ theme }) => theme.MIXINS.flexBox('column', 'center', 'center')};
-    padding: 16px 0px;
-    width: 100%;
-    background-color: ${({ theme }) => theme.colors.gray['747474']};
-    color: ${({ theme }) => theme.colors.white};
-    transition: 0.3s;
-    opacity: .75;
-    border: none;
-    cursor: pointer;
-  
-    &:hover {
-      opacity: 1;
-      background-color: ${({ theme }) => theme.colors.black};
-    }
-  `,
-};
+import * as S from './page.style';
 
 interface FormProps {
   title: string;
