+++ src/app/page.tsx
@@ -5,8 +5,8 @@ import React, { useEffect, useState } from 'react';
 import { useSearchParams } from 'next/navigation';
 import * as S from './page.style';
 import { postAPI } from '@/api/post';
-import { TagCard } from '@/components/card/TagCard';
-import { PostCard } from '@/components/card/PostCard';
+import { TagCard } from '@/components/card/TagCard/TagCard';
+import { PostCard } from '@/components/card/PostCard/PostCard';
 import { LoadingSkeleton } from '@/components/loading/LoadingSkeleton';
 import { Pagination } from '@/components/Pagination';
 import { Post } from '@/types/Post';
