+++ src/components/card/TagCard/TagCard.tsx
@@ -1,0 +1,9 @@
+import * as S from './TagCard.style';
+
+interface Props {
+  tagName: string;
+}
+
+export function TagCard({ tagName }: Props) {
+  return <S.cardContainer>{`#${tagName}`}</S.cardContainer>;
+}
