+++ src/components/layouts/Main/Main.style.ts
@@ -1,0 +1,6 @@
+import { styled } from 'styled-components';
+
+export const container = styled.main`
+  ${({ theme }) => theme.MIXINS.flexBox('column')}
+  width: 100%;
+`;
